requirements.txt
# Recommended runtime: Python 3.11 or 3.12 on a clean machine/venv.
# Python 3.9 may require native build tools for some transitive dependencies.
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.6
Pillow==10.1.0
stripe==5.0.0
torch>=2.2.0
torchvision>=0.17.0
